+++ src/char/int_pet.c
@@ -637,7 +637,7 @@ int mapif_load_pet(int fd,int account_id,int char_id,int pet_id)
 int mapif_save_pet(int fd,int account_id,struct s_pet *data)
 {
 	if(sizeof(struct s_pet) != RFIFOW(fd,2) - 8) {
-		printf("inter pet: data size error %d %d\n",sizeof(struct s_pet),RFIFOW(fd,2)-8);
+		printf("inter pet: data size error %lu %d\n",(unsigned long)sizeof(struct s_pet),RFIFOW(fd,2)-8);
 		return 0;
 	}
