+++ src/common/grfio.h
@@ -28,8 +28,8 @@ void* grfio_read(const char*);	// GRFIO data file read
 void* grfio_reads(const char*,int*);	// GRFIO data file read & size get
 int grfio_size(const char*);		// GRFIO data file size get
 
-int decode_zip(char *dest, unsigned long* destLen, const char* source, unsigned long sourceLen);
-int encode_zip(char *dest, unsigned long* destLen, const char* source, unsigned long sourceLen);
+int decode_zip(char *dest, unsigned long* destLen, const char* source, unsigned int sourceLen);
+int encode_zip(char *dest, unsigned long* destLen, const char* source, unsigned int sourceLen);
 
 void grfio_load_zlib(void);
 unsigned long grfio_crc32(const unsigned char *buf, unsigned int len);
